requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.13.0
pydantic==2.5.0
llama-index==0.9.13
llama-index-llms-gemini==0.1.6
openinference-instrumentation-llama_index==4.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==8.3.5
pytest-asyncio=1.0.0
asgi-lifespan==2.1.0
arize-phoenix-otel==0.10.1